Width/height of images should be integers

+++ django/views.py
@@ -51,7 +51,7 @@ def index(request):
                     if onlyerror:
                         item_info[short_path] = open(item).read()
                     continue
-                dimensions = [x*strip_decode.zoom for x in Image.open(item).size]
+                dimensions = [int(x*strip_decode.zoom) for x in Image.open(item).size]
                 item_info[short_path] = {"width":dimensions[0], "height":dimensions[1]}
             if len(item_info) > 0:
                 strips[strip] = {"onlyerror":onlyerror, "homepage": strip_decode.homepage, "items": sorted(item_info.items()), "desc": strip_decode.desc}
